HARVESTER: move harvester CRDs into virtual directory
